x86Emitter: Fix incorrect displacement for some 64-bit values

+++ common/emitter/x86types.h
@@ -75,7 +75,7 @@ namespace x86Emitter
 	template <typename T>
 	static __fi bool is_s8(T imm)
 	{
-		return (s8)imm == (s32)imm;
+		return (s8)imm == (s64)imm;
 	}
 
 	template <typename T>
